Write byte array in chunks if needed
On DMD64, the length of an array is stored in a ulong,
which is larger than a uint. Since `PHYSFS_write` and
`PHYSFS_read` expect a uint, this fails to compile.

The fix is easy - just check at compile-time whether
the types are mismatched, and write in chunks if we have to.